Fix backspace when editing a multiline cmdline
Once the cmdline had passed the width of the screen, adding additional
characters introduced a spurious newline, and another newline at the
width of input.  Furthermore, hitting backspace would not start
redrawing at the end of input, but rather at the beginning of the
current line - resulting in extra duplicate lines scrolling the console.

First, fix the assumption that the length of cmdline is the width - it
needs to include the length of the prompt (i.e., length of input and
space).

Second, fix the behavior of single-line redraw (i.e., redraw == 1) to
move the cursor to the row the line begins at.

Third, don't scroll the cursor down when a line wrap would occur - it's
not necessary since line wrap is enabled, and results in the extra blank
line.

Finally, comment all used escape sequences so that I don't need to look
them up again.

Signed-off-by: Robbie Harwood <rharwood@redhat.com>